+++ src/MemcachedTiny/IMemcachedClientAsync.cs
@@ -28,7 +28,7 @@ namespace MemcachedTiny
         /// <param name="bytes">缓存数据</param>
         /// <returns>一个代表异步任务的<see cref="Task" /></returns>
         /// <param name="cancellation">（可选）一个标识操作取消的<see cref="CancellationToken"/></param>
-        Task<IResult> SetAsync(string key, int flags, int second, byte[] bytes, CancellationToken cancellation = default);
+        Task<IResult> SetAsync(string key, int flags, uint second, byte[] bytes, CancellationToken cancellation = default);
 
         /// <summary>
         /// 获取一个缓存值（如果存在的话）
@@ -46,7 +46,7 @@ namespace MemcachedTiny
         /// <param name="cas">数据版本</param>
         /// <param name="cancellation">（可选）一个标识操作取消的<see cref="CancellationToken"/></param>
         /// <returns>一个代表异步任务的<see cref="Task" /></returns>
-        Task<IResult> TouchAsync(string key, uint second, uint cas, CancellationToken cancellation = default);
+        Task<IResult> TouchAsync(string key, uint second, long cas, CancellationToken cancellation = default);
 
         /// <summary>
         /// 获取一个缓存值并更新缓存时间（如果存在的话）
